TaglistのリファクタリングおよびDraftボタンの調整
