+++ astro/src/components/Client/bsky/buttons/DraftSaveButton.tsx
@@ -4,13 +4,10 @@ import { Dispatch, SetStateAction } from "react"
 // components
 import ProcButton from "../../common/ProcButton"
 
-// backend api
-import { richTextFacetParser } from "@/utils/richTextParser"
-
 // service
-import { setSavedTags, readSavedTags, saveDrafts, readDrafts } from "@/utils/useLocalStorage";
+import { saveDrafts, readDrafts } from "@/utils/useLocalStorage";
 import { msgInfo } from "../../common/types"
-
+import saveTagToSavedTags from "../lib/saveTagList";
 
 export const Component = ({
     postText,
@@ -19,7 +16,6 @@ export const Component = ({
     isProcessing,
     setProcessing,
     setMsgInfo,
-    disabled,
 }: {
     postText: string,
     setPostText: Dispatch<SetStateAction<string>>,
@@ -27,10 +23,14 @@ export const Component = ({
     isProcessing: boolean,
     setProcessing: Dispatch<SetStateAction<boolean>>,
     setMsgInfo: Dispatch<SetStateAction<msgInfo>>,
-    disabled: boolean
 }) => {
-    // 保存できるタグの上限
-    const maxTagCount = 8
+
+    /**
+     * 下書きとして保存可能かどうかを判定します
+     * @returns ポスト可能な場合true
+     */
+    const isValidDraft = () => postText.length >= 1
+
 
     const handleSavePost = () => {
         const isValidPost = (): boolean => {
@@ -50,41 +50,9 @@ export const Component = ({
             setProcessing(true)
             // Postを押したら強制的にフォーカスアウトイベントを発火
         }
-        
+
         initializePost()
-
-        try {
-            // Hashtagが含まれている場合はブラウザに保存
-            const richTextLinkParser = new richTextFacetParser("tag")
-            const parseResult = richTextLinkParser.getFacet(postText)
-
-            const savedTag = readSavedTags()
-            
-            let taglist: string[] = (savedTag !== null) ? savedTag : []
-            parseResult.forEach((value) => {
-                const tagName = value
-                const tagIndex = taglist.indexOf(tagName)
-                if (tagIndex < 0) {
-                    // タグが存在しない場合は先頭に追加
-                    taglist.unshift(tagName)
-                } else {
-                    // タグが存在する場合は先頭に移動
-                    taglist.splice(tagIndex, 1)
-                    taglist.unshift(tagName)
-                }
-            })
-
-            setSavedTags(taglist.slice(0, maxTagCount))
-        } catch (error: unknown) {
-            let msg: string = "Unexpected Unknown Error"
-            if (error instanceof Error) {
-                msg = error.name + ": " + error.message
-            }
-            setMsgInfo({
-                msg: msg,
-                isError: true
-            })
-        }
+        saveTagToSavedTags({ postText })
 
         const newDrafts = [postText, ...readDrafts() || []]
 
@@ -92,12 +60,12 @@ export const Component = ({
         setDrafts(newDrafts)
 
         setPostText("")
-        
+
         setProcessing(false)
 
         setMsgInfo({
-          msg: "下書きを保存しました。",
-          isError: false
+            msg: "下書きを保存しました。",
+            isError: false
         })
     }
 
@@ -113,7 +81,7 @@ export const Component = ({
             </span>}
             color="blue"
             className={["my-0", "py-0.5", "align-middle"]}
-            disabled={disabled} />
+            disabled={!isValidDraft()} />
     )
 }
 export default Component
